requirements.txt
certifi==2018.11.29
chardet==3.0.4
confusable-homoglyphs==3.1.1
dj-database-url==0.5.0
Django==1.11
django-annoying==0.10.4
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-mathfilters==0.4.0
django-passwords==0.3.12
django-registration==2.4.1
django-stringfield==0.3.0
gunicorn==19.9.0
idna==2.8
Pillow==5.2.0
psycopg2==2.7.7
psycopg2-binary==2.7.6.1
python-dateutil==2.7.5
python-decouple==3.1
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
whitenoise==3.3.1
